requirements.txt
flask==2.0.1
torch==2.2.0
torchvision==0.17.0
pillow==10.0.0
werkzeug==2.0.1
gunicorn==20.1.0
